Move ban management out of permissionmanager.
